sync revision 1647 	- added simple workaround for coherent memory on Windows 7 	- use Metal resource heaps 	- editor application stub 		* builds natively on macOS 		* builds with GNUstep on *nix (requires clang for Obj-C 		  2.0) 		* Windows currently not supported

+++ Engine/Asset/Font.c
@@ -60,7 +60,7 @@ E_LoadFontAsset(struct Stream *stm, struct Font *fnt)
 	
 	texSize = Re_deviceInfo.limits.maxTextureSize;
 	if (texSize < hdr.texSize) {
-		uint32_t curSize = hdr.texSize;
+		uint64_t curSize = hdr.texSize;
 
 		while (curSize > texSize) {
 			E_StreamSeek(stm, curSize * curSize, IO_SEEK_CUR);
